src/main/java/com/controller/TransferController.java
package com.controller;

import com.domain.Transfer;
import com.service.TransferService;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;

import javax.inject.Inject;

@Controller
@RequestMapping("/transfers")
public class TransferController {

    private TransferService transferService;

    @Inject
    public TransferController(TransferService transferService) {
        this.transferService = transferService;
    }

    @RequestMapping(method = RequestMethod.POST)
    @ResponseStatus(HttpStatus.CREATED)
    @ResponseBody
    public Transfer createTransfer(@RequestBody Transfer transfer) {
        transferService.createTransfer(transfer);
        return transfer;
    }

    @RequestMapping(value = "/{id}", method = RequestMethod.GET)
    @ResponseStatus(HttpStatus.OK)
    @ResponseBody
    public Transfer getTransfer(@PathVariable Long id) {
        return transferService.getTransfer(id);
    }
}
